upd authorization

+++ PlaneMarket/PlaneMarket_back/api/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import generics, status, viewsets
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 from rest_framework_simplejwt.tokens import RefreshToken
@@ -20,6 +20,9 @@ class LogoutView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         try:
             # getting refresh token from data request
             refresh_token = request.data.get('refresh_token')
@@ -32,7 +35,6 @@ class LogoutView(APIView):
 
 
 class CurrentUserView(APIView):
-    permission_classes = [IsAuthenticated]
 
     def get(self, request):
         # Возвращаем информацию о текущем аутентифицированном пользователе
@@ -45,12 +47,16 @@ class CurrentUserView(APIView):
     
 @api_view(['GET', 'POST'])
 def manufacturers_list(request):
+
     if request.method == 'GET':
         manufacturers = Manufacturer.objects.all()
         serializer = ManufacturerSerializer(manufacturers, many=True)
         return Response(serializer.data)
 
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = ManufacturerSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -58,6 +64,7 @@ def manufacturers_list(request):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticated])
 def manufacturer_detail(request, id):
     manufacturer = Manufacturer.objects.get(id=id)
 
@@ -66,6 +73,9 @@ def manufacturer_detail(request, id):
         return Response(serializer.data)
 
     elif request.method == 'PUT':
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = ManufacturerSerializer(manufacturer, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -73,10 +83,14 @@ def manufacturer_detail(request, id):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         manufacturer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 @api_view(['GET'])
+# @permission_classes([IsAuthenticated])
 def planes_by_manufacturer(request, id):
     planes = Plane.objects.filter(manufacturer_id=id)
     if not planes:
@@ -94,6 +108,9 @@ class PlaneListAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = PlaneSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -105,15 +122,14 @@ class PlaneDetailAPIView(APIView):
     #permission_classes = [IsAuthenticated]
 
     def get(self, request, id):
-        try:
-            plane = get_object_or_404(Plane, id=id)
-        except Plane.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
+        plane = get_object_or_404(Plane, id=id)
         serializer = PlaneSerializer(plane)
         return Response(serializer.data)
 
     def put(self, request, id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         plane = Plane.objects.get(id=id)
         serializer = PlaneSerializer(plane, data=request.data)
         if serializer.is_valid():
@@ -122,12 +138,17 @@ class PlaneDetailAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
+        if not request.user.is_authenticated:
+            return Response({"detail": "Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
+
         plane = Plane.objects.get(id=id)
         plane.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class TopTenPlaneListAPIView(APIView):
+    # permission_classes = [IsAuthenticated]
+
     def get(self, request):
         planes = Plane.objects.order_by('-price')[:10]
         if not planes:
@@ -139,14 +160,20 @@ class TopTenPlaneListAPIView(APIView):
 
 class PlaneViewSet(viewsets.ModelViewSet):
     #permission_classes = [IsAuthenticated]
+
     serializer_class = PlaneSerializer
     queryset = Plane.objects.all()
 
     def perform_create(self, serializer):
-        serializer.save(manufacturer=self.request.user.manufacturer)
+        manufacturer = getattr(self.request.user, 'manufacturer', None)
+        if manufacturer:
+            serializer.save(manufacturer=manufacturer)
+        else:
+            serializer.save()
 
 
 @api_view(['GET'])
+# @permission_classes([IsAuthenticated])
 def orders_by_plane(request, id):
     orders = Order.objects.filter(plane_id=id)
     if not orders:
@@ -156,6 +183,7 @@ def orders_by_plane(request, id):
 
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def customer_orders(request, id):
     orders = Order.objects.filter(customer_id=id)
     if not orders:
@@ -165,6 +193,7 @@ def customer_orders(request, id):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def create_order_for_authenticated_user(request):
     customer = getattr(request.user, 'customer', None)
     if not customer:
@@ -183,23 +212,3 @@ def create_order_for_authenticated_user(request):
     serializer = OrderSerializer(order)
     return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-
-# # new endpoint
-#
-# # Логин — получение JWT токенов
-# class CustomTokenObtainPairView(TokenObtainPairView):
-#     pass
-#
-# # Логаут — аннулирование refresh токена
-# class LogoutView(APIView):
-#     permission_classes = [IsAuthenticated]
-#
-#     def post(self, request):
-#         try:
-#             # Получаем refresh токен из данных запроса
-#             refresh_token = request.data.get('refresh_token')
-#             token = RefreshToken(refresh_token)
-#             token.blacklist()  # Добавляем в черный список, чтобы токен больше не был действителен
-#             return Response({"detail": "Logged out successfully"}, status=200)
-#         except Exception as e:
-#             return Response({"error": "Something went wrong"}, status=400)
